refactor: moved message to its own component also other minor refactors
