+++ src/client/components/App.tsx
@@ -4,7 +4,7 @@ import {TextMessage, CreateRoomMessage, JoinRoomMessage, Message} from 'fluxxcha
 import {get} from 'lodash';
 import Menu from './Menu';
 import ChatRoom from '../scenes/ChatRoom';
-import NavigationBar from '../scenes/NavigationBar/navbar';
+import NavigationBar from './NavBar';
 import '../styles.css';
 
 interface State {
