+++ src/client/components/NavBar.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../../styles.css';
+import '../styles.css';
 
 class NavigationBar extends React.Component {
 	public render() {
